docker and tests

+++ src/main/java/CourseApp/CourseApp/Service/CourseService.java
@@ -1,27 +1,42 @@
-package CourseApp.CourseApp;
+package CourseApp.CourseApp.Service;
 
+import CourseApp.CourseApp.CourseRecommender;
+import CourseApp.CourseApp.DTO.AuthorDto;
+import CourseApp.CourseApp.DTO.CourseDto;
+import CourseApp.CourseApp.Entity.Course;
+import CourseApp.CourseApp.Mapper.AuthorMapper;
+import CourseApp.CourseApp.Mapper.CourseMapper;
+import CourseApp.CourseApp.Repository.AuthorRepository;
+import CourseApp.CourseApp.Repository.CourseRepository;
+import CourseApp.CourseApp.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.HashMap;
 import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
 
 @Service
 public class CourseService {
+    private final AuthorMapper authorMapper;
+    private final AuthorRepository authorRepository;
     private final CourseRepository courseRepository;
     private final CourseMapper courseMapper;
     CourseRecommender recommender;
-    private HashMap<Integer,Course>courses;
+
     private final AtomicInteger idGenerator = new AtomicInteger();
     @Autowired //Here precedence is given using Primary
-    public CourseService(CourseRepository courseRepository, CourseMapper courseMapper, CourseRecommender recommender) {
+    public CourseService(AuthorMapper authorMapper, AuthorRepository authorRepository, CourseRepository courseRepository, CourseMapper courseMapper) {
+        this.authorMapper = authorMapper;
+        this.authorRepository = authorRepository;
         this.courseRepository = courseRepository;
         this.courseMapper = courseMapper;
         this.recommender = recommender;
-        courses = new HashMap<>();
+
 
     }
 
@@ -34,6 +49,12 @@ public class CourseService {
 
     public List<Course> recommendCourses() {
         return recommender.recommendCourses();
+    }
+
+    @Transactional
+    public Page<CourseDto> getAllCourses(Pageable pageable) {
+        Page<Course> courses = courseRepository.findAll(pageable);
+        return courseMapper.coursePageToCourseDTOPage(courses, pageable);
     }
 
     public CourseDto getCourseById(Long id) {
@@ -56,8 +77,17 @@ public class CourseService {
         return courseMapper.toDto(courseRepository.save(course));
     }
 
-    public void deleteCourse(Long id) {
-        courseRepository.deleteById(id);
+    public boolean deleteCourse(Long id) {
+        if (courseRepository.existsById(id)) {
+            courseRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
+    }
+    public Optional<AuthorDto> getAuthorByEmail(String email) {
+        return authorRepository.findByEmail(email)
+                .map(authorMapper::toDto);
     }
 
 //    public Course addCourse(Course course){
